Use context functions + split IrcBot + fix upsert
Context functions provide a better way to do DI. It's like a ReaderMonad
but implicitly managed by the compiler.

The IrcBot was not testable, hence I extracted the logic into a port
TextBot.

MySQL / MariaDB affected rows count is 0 if the row was not updated
after an upsert. I should have been able to use `useAffectedRows` flag
to fix this, but somehowe it didn't work, so I used a new column with a
random value, only to have something to update and get my updated's row
id back.
See: https://bugs.mysql.com/bug.php?id=39352